+++ CHANGES.rst
@@ -4,12 +4,13 @@ Changelog
 Here you can see the full list of changes between each WTForms-Alchemy release.
 
 
-0.15.0 (2016-01-xx)
+0.15.0 (2016-01-27)
 ^^^^^^^^^^^^^^^^^^^
 
 - Moved GroupedQuerySelectField from WTForms-Components package to WTForms-Alchemy
 - Moved WeekdaysField from WTForms-Components package to WTForms-Alchemy
 - Moved PhoneNumberField from WTForms-Components package to WTForms-Alchemy
+- Moved Unique validator from WTForms-Components package to WTForms-Alchemy
 
 
 0.14.0 (2016-01-23)